update WS handler
