+++ application.go
@@ -99,11 +99,21 @@ func (app *Application) cast(streamID uuid.UUID, pck av.Packet) error {
 	return nil
 }
 
-func (app *Application) ext(streamID uuid.UUID) bool {
+func (app *Application) exists(streamID uuid.UUID) bool {
 	app.Streams.Lock()
 	defer app.Streams.Unlock()
 	_, ok := app.Streams.Streams[streamID]
 	return ok
+
+}
+
+func (app *Application) existsWithType(streamID uuid.UUID, streamType string) bool {
+	app.Streams.Lock()
+	defer app.Streams.Unlock()
+	stream, ok := app.Streams.Streams[streamID]
+	supportedTypes := stream.SupportedStreamTypes
+	typeEnabled := typeExists(streamType, supportedTypes)
+	return ok && typeEnabled
 }
 
 func (app *Application) codecAdd(streamID uuid.UUID, codecs []av.CodecData) {
